+++ src/sheetforge/__init__.py
@@ -15,6 +15,7 @@ from sheetforge.formulas import (
     FormulaTranslationDiagnostic,
     translate_formula_cell,
 )
+from sheetforge.formulas_oracle import FormulasWorkbookOracle
 from sheetforge.generation import (
     GeneratedModuleContract,
     GeneratedSymbol,
@@ -28,6 +29,15 @@ from sheetforge.graph import (
     DependencyGraph,
     build_dependency_graph,
 )
+from sheetforge.oracles import (
+    OracleDiagnostic,
+    OracleOutputs,
+    OracleRequest,
+    OracleResult,
+    WorkbookOracle,
+    missing_optional_dependency_diagnostic,
+)
+from sheetforge.oracle_validation import build_oracle_validation_report
 from sheetforge.references import (
     WorkbookReference,
     normalize_cell_reference,
@@ -62,6 +72,7 @@ __all__ = [
     "FormulaExpressionNode",
     "FormulaRecord",
     "FormulaTranslationDiagnostic",
+    "FormulasWorkbookOracle",
     "GeneratedModuleContract",
     "GeneratedSymbol",
     "GenerationDiagnostic",
@@ -69,20 +80,27 @@ __all__ = [
     "MISSING_VALUE",
     "NamedRangeRecord",
     "OracleConfig",
+    "OracleDiagnostic",
+    "OracleOutputs",
+    "OracleRequest",
+    "OracleResult",
     "ScenarioInput",
     "ScenarioOutput",
     "SheetRecord",
     "ValidationReport",
     "ValidationScenario",
+    "WorkbookOracle",
     "WorkbookReference",
     "WorkbookRecord",
     "__version__",
     "build_dependency_graph",
+    "build_oracle_validation_report",
     "build_validation_report",
     "compare_scalar_output",
     "extract_workbook",
     "generate_python_module",
     "load_validation_scenario",
+    "missing_optional_dependency_diagnostic",
     "normalize_cell_reference",
     "normalize_reference",
     "symbol_name_for_cell_ref",